+++ public/constants.jsx
@@ -1,0 +1,3 @@
+export const COLORS = {
+    base: '#FBC846'
+};
