+++ styles/Home.module.scss
@@ -1,0 +1,92 @@
+$base-color: #FBC846;
+
+.container {
+  padding: 0 4rem;
+}
+
+.main {
+  min-height: 100vh;
+  padding: 4rem 0;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.title a {
+  color: #0070f3;
+  text-decoration: none;
+}
+
+.title a:hover,
+.title a:focus,
+.title a:active {
+  text-decoration: underline;
+}
+
+.title {
+  margin: 0;
+  line-height: 1.15;
+  font-size: 4rem;
+}
+
+.title {
+  text-align: center;
+}
+
+.componentTitle {
+  padding: 10px 0;
+  font-size: 20px;
+  font-weight: 600;
+}
+
+.headingIcon {
+  font-size: 20px; 
+  width: 20px; 
+  height: 20px; 
+  padding-top: 30px;
+  padding-bottom: 30px; 
+}
+
+.gridContainer {
+  display: grid;
+}
+
+.contentBox {
+  width: min(100%, 600px);
+  border-style: solid;
+  border-width: 1px;
+  div {
+    padding-top: 20px; 
+  }
+  .sub {
+    margin-bottom: 20px; 
+  }
+  text-align: center;
+  font-size: 20px; 
+}
+
+@media only screen and (max-width: 600px) {
+  .gridContainer {
+    text-align: center;
+    grid-template-columns: auto;
+  }
+
+  .secondItem {
+    grid-row-start: 1;
+    grid-row-end: 3;
+    background-image: url('../public/pictures/meo.jpg');
+  }
+}
+
+@media only screen and (min-width: 600px) {
+  .gridContainer {
+    text-align: center;
+    grid-template-columns: auto auto;
+  }
+
+  .secondItem {
+    vertical-align: middle;
+  }
+}
