+++ styles/globals.scss
@@ -11,4 +11,6 @@ a {
 
 * {
   box-sizing: border-box;
+  font-family: 'Lora';
 }
+
